requirements.txt
Flask==2.3.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.3
click==8.1.6
itsdangerous==2.1.2
pyodbc==4.0.35
pymongo==4.4.1
azure-storage-blob
gunicorn==22.0.0
python-dotenv
schedule